Added support for passing userInfo from TemplateRenderer to TemplateDataEncoder to Encoder
These changes allow for data to be passed through the TemplateRenderer all the way through to the Encoder that is passed to the view model objects during encoding. This allows for passing request-level information such as locale to the view model objects for selecting formatting, language, etc. during view model encoding.

Existing public APIs (TemplateDataEncoder.encode) were augmented with default values for the new argument to maintain backwards API compatibility, while providing for the new functionality.